refactor: move products routes, schemas, and service to products/
- api/routes/products.py → products/routes.py
- schemas/product.py → products/schemas.py
- services/product_service.py → products/service.py
- Update imports in routes and test patches
