+++ app/main.py
@@ -17,7 +17,8 @@ from app.auth.routes import router as auth_router
 from app.orgs.routes import router as orgs_router
 from app.admin.routes_orgs import router as admin_orgs_router
 from app.admin.routes_users import router as admin_users_router
-from app.api.routes import health, suppliers, products
+from app.api.routes import health, suppliers
+from app.products.routes import router as products_router
 
 # ── Structured JSON logging to stdout → Docker → CloudWatch ──
 handler = logging.StreamHandler(sys.stdout)
@@ -78,7 +79,7 @@ def create_app() -> FastAPI:
     app.include_router(admin_users_router)
     app.include_router(orgs_router)
     app.include_router(suppliers.router)
-    app.include_router(products.router)
+    app.include_router(products_router)
 
     return app
